# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Wisata Indonesia API - Tourism Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cities:
    get:
      operationId: getCities
      tags: [cities]
      summary: List all cities
      parameters:
        - name: provinceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
    post:
      operationId: createCity
      tags: [cities]
      summary: Create a city
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
  /cities/{id}:
    put:
      operationId: updateCity
      tags: [cities]
      summary: Update a city
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCityInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
    delete:
      operationId: deleteCity
      tags: [cities]
      summary: Delete a city
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /blog:
    get:
      operationId: getBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: published
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /banners:
    get:
      operationId: getBanners
      tags: [banners]
      summary: List banners
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    put:
      operationId: updateBanner
      tags: [banners]
      summary: Update a banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/custom-trips:
    get:
      operationId: getCustomTrips
      tags: [admin]
      summary: List custom trip requests
      responses:
        "200":
          description: List of custom trip requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomTrip"
    post:
      operationId: createCustomTrip
      tags: [admin]
      summary: Submit a custom trip request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomTripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomTrip"
  /admin/custom-trips/{id}:
    put:
      operationId: updateCustomTrip
      tags: [admin]
      summary: Update custom trip status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomTripInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomTrip"
components:
  schemas:
...
    Province:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - createdAt
...
    City:
      type: object
      properties:
        id:
          type: integer
        provinceId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        province:
          $ref: "#/components/schemas/Province"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - provinceId
        - name
        - slug
        - createdAt
    CreateCityInput:
      type: object
      properties:
        provinceId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required:
        - provinceId
        - name
        - slug
...
    TourPackage:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        duration:
          type: integer
        maxParticipants:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        includes:
          type: array
          items:
            type: string
        excludes:
          type: array
          items:
            type: string
        itinerary:
          type: string
          nullable: true
        provinceId:
          type: integer
          nullable: true
        cityId:
          type: integer
          nullable: true
        isFeatured:
          type: boolean
        isActive:
          type: boolean
        rating:
          type: number
          nullable: true
        totalReviews:
          type: integer
        province:
          $ref: "#/components/schemas/Province"
          nullable: true
        city:
          $ref: "#/components/schemas/City"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - slug
        - price
        - duration
        - images
        - includes
        - excludes
        - isFeatured
        - isActive
        - totalReviews
        - createdAt
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        brand:
          type: string
          nullable: true
        capacity:
          type: integer
        pricePerDay:
          type: number
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - capacity
        - pricePerDay
        - images
        - features
        - isAvailable
        - createdAt
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        status:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        packageId:
          type: integer
          nullable: true
        vehicleId:
          type: integer
          nullable: true
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        participants:
          type: integer
          nullable: true
        totalPrice:
          type: number
        notes:
          type: string
          nullable: true
        package:
          $ref: "#/components/schemas/TourPackage"
          nullable: true
        vehicle:
          $ref: "#/components/schemas/Vehicle"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - type
        - status
        - customerName
        - customerEmail
        - startDate
        - totalPrice
        - createdAt
    CreateBookingInput:
      type: object
      properties:
        type:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        packageId:
          type: integer
          nullable: true
        vehicleId:
          type: integer
          nullable: true
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        participants:
          type: integer
          nullable: true
        totalPrice:
          type: number
        notes:
          type: string
          nullable: true
      required:
        - type
        - customerName
        - customerEmail
        - startDate
        - totalPrice
    UpdateBookingInput:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - status
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
          nullable: true
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        authorName:
          type: string
        isPublished:
          type: boolean
        publishedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - slug
        - content
        - authorName
        - tags
        - isPublished
        - createdAt
    CreateBlogPostInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
          nullable: true
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        authorName:
          type: string
        isPublished:
          type: boolean
      required:
        - title
        - slug
        - content
        - authorName
        - isPublished
...
    Banner:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        imageUrl:
          type: string
        ctaText:
          type: string
          nullable: true
        ctaLink:
          type: string
          nullable: true
        isActive:
          type: boolean
        order:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - imageUrl
        - isActive
        - order
        - createdAt
    CreateBannerInput:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        imageUrl:
          type: string
        ctaText:
          type: string
          nullable: true
        ctaLink:
          type: string
          nullable: true
        isActive:
          type: boolean
        order:
          type: integer
      required:
        - title
        - imageUrl
        - isActive
        - order
...
    CustomTrip:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        participants:
          type: integer
        budget:
          type: number
          nullable: true
        requirements:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerEmail
        - destination
        - startDate
        - participants
        - status
        - createdAt
    CreateCustomTripInput:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        participants:
          type: integer
        budget:
          type: number
          nullable: true
        requirements:
          type: string
          nullable: true
      required:
        - customerName
        - customerEmail
        - destination
        - startDate
        - participants
    UpdateCustomTripInput:
      type: object
      properties:
        status:
          type: string
      required:
        - status