# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wisata Indonesia API - Tourism Platform
 * OpenAPI spec version: 0.1.0
 */
import type { TourPackage } from "./tourPackage";
import type { Vehicle } from "./vehicle";

export interface Booking {
  id: number;
  type: string;
  status: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  packageId?: number | null;
  vehicleId?: number | null;
  startDate: string;
  endDate?: string | null;
  participants?: number | null;
  totalPrice: number;
  notes?: string | null;
  package?: TourPackage | null;
  vehicle?: Vehicle | null;
  createdAt: string;
}