# lib/api-zod/src/generated/types/customTrip.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wisata Indonesia API - Tourism Platform
 * OpenAPI spec version: 0.1.0
 */

export interface CustomTrip {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  destination: string;
  startDate: string;
  endDate?: string | null;
  participants: number;
  budget?: number | null;
  requirements?: string | null;
  status: string;
  createdAt: string;
}